requirements.txt
aiogram==3.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
aiohttp==3.8.5
